Move quotes to own screen
